+++ stories/DataDisplay.stories.ts
@@ -1,0 +1,11 @@
+import DataDisplay from './DataDisplay.svelte';
+
+export default {
+  title: 'Design System/Components/Data Display',
+  component: DataDisplay,
+  parameters: {
+    layout: 'fullscreen'
+  }
+};
+
+export const States = {};